+++ disseminator.py
@@ -6,7 +6,7 @@ Call custom workflows to retrieve work-related files and metadata
 and upload them in the appropriate format to various platforms.
 """
 
-__version__ = '0.1.14'
+__version__ = '0.1.15'
 
 import argparse
 import logging
@@ -20,6 +20,10 @@ from culuploader import CULUploader
 from crossrefuploader import CrossrefUploader
 from fsuploader import FigshareUploader
 from zenodouploader import ZenodoUploader
+from museuploader import MUSEUploader
+from jstoruploader import JSTORUploader
+from ebscouploader import EBSCOUploader
+from proquestuploader import ProquestUploader
 
 UPLOADERS = {
     "InternetArchive": IAUploader,
@@ -29,6 +33,10 @@ UPLOADERS = {
     "Crossref": CrossrefUploader,
     "Figshare": FigshareUploader,
     "Zenodo": ZenodoUploader,
+    "ProjectMUSE": MUSEUploader,
+    "JSTOR": JSTORUploader,
+    "EBSCOHost": EBSCOUploader,
+    "ProQuest": ProquestUploader,
 }
 
 UPLOADERS_STR = ', '.join("%s" % (key) for (key, _) in UPLOADERS.items())